+++ generators/csharp/base/src/asIs/test/RawClientTests/IdempotentHeadersTests.Template.cs
@@ -144,13 +144,13 @@ public class IdempotentHeadersTests
     }
 
     [Test]
-    public void CheckForIdempotencyHeadersSupport()
+    public async SystemTask CheckForIdempotencyHeadersSupport()
     {
         _server
             .Given(WireMockRequest.Create().WithPath("/test").WithParam("foo", "bar").UsingGet())
             .RespondWith(WireMockResponse.Create().WithStatusCode(200).WithBody("Success"));
 
-        var request = _rawClient.CreateHttpRequest(new JsonRequest()
+        var request = await _rawClient.CreateHttpRequestAsync(new JsonRequest()
         {
             BaseUrl = _baseUrl,
             Method = HttpMethod.Get,
@@ -162,7 +162,7 @@ public class IdempotentHeadersTests
                 IdempotencyExpiration = 3600
             }
             
-        });
+        }).ConfigureAwait(false);
 
         request.Headers.TryGetValues("IDEMPOTENCY-KEY", out var idempotencyKey);
         Assert.That(idempotencyKey?.First(), Is.EqualTo("1234567890"));