+++ seed/csharp-sdk/oauth-client-credentials-mandatory-auth/src/SeedOauthClientCredentialsMandatoryAuth/SeedOauthClientCredentialsMandatoryAuthClient.cs
@@ -35,9 +35,10 @@ public partial class SeedOauthClientCredentialsMandatoryAuthClient
             clientSecret,
             new AuthClient(new RawClient(clientOptions.Clone()))
         );
-        clientOptions.Headers["Authorization"] = new Func<string>(() =>
-            tokenProvider.GetAccessTokenAsync().Result
-        );
+        clientOptions.Headers["Authorization"] =
+            new Func<global::System.Threading.Tasks.ValueTask<string>>(async () =>
+                await tokenProvider.GetAccessTokenAsync().ConfigureAwait(false)
+            );
         _client = new RawClient(clientOptions);
         Auth = new AuthClient(_client);
         Nested = new NestedClient(_client);